Extracted custom exceptions into categorized modules: `Arguments.py`, `Initializations.py`, and `Misc.py`, for improved organization and clarity. Updated `__init__.py` to reflect new structure.
